separated the calculation function for nights and total payment, and added suggestion for 1 map instead of multiple ones in lambdarepository
